requirements.txt
torch==0.3.1
torchvision
torchtext==0.2.3
fastai==0.7.0
imutils
opencv-python
scipy
numpy
python-chess
Flask